repo相关的所有api完成，可供测试
新增了restock和获取restock历史

+++ Controllers/RepoController.cs
@@ -5,6 +5,7 @@ using Elderly_Canteen.Services.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.Security.Claims;
+using System.Security.Principal;
 
 namespace Elderly_Canteen.Controllers
 {
@@ -20,9 +21,17 @@ namespace Elderly_Canteen.Controllers
         }
 
         [HttpGet("search")]
-        public async Task<IActionResult> GetAllIngredient()
+        public async Task<IActionResult> GetAllIngredient(string? name)
         {
-            var response = await _repoService.GetRepo();
+            var identity = User.FindFirst(ClaimTypes.Role)?.Value;
+            if (identity != "admin")
+                return BadRequest(new RestockResponseDto
+                {
+                    Success = false,
+                    Message = "you are not administrator",
+                    Data = null,
+                });
+            var response = await _repoService.GetRepo(name);
             if (response == null)
             {
                 return NotFound();
@@ -56,6 +65,14 @@ namespace Elderly_Canteen.Controllers
         [HttpPut("update")]
         public async Task<ActionResult> UpdateRepo(RepoRequestDto dto)
         {
+            var identity = User.FindFirst(ClaimTypes.Role)?.Value;
+            if (identity != "admin")
+                return BadRequest(new RestockResponseDto
+                {
+                    Success = false,
+                    Message = "you are not administrator",
+                    Data = null,
+                });
             if (dto == null)
             {
                 return BadRequest(new IngreResponseDto
@@ -78,6 +95,14 @@ namespace Elderly_Canteen.Controllers
         [HttpDelete("delete/{id}/{expiry}")]
         public async Task<ActionResult> DeleteRepo(string id, DateTime expiry)
         {
+            var identity = User.FindFirst(ClaimTypes.Role)?.Value;
+            if (identity != "admin")
+                return BadRequest(new RestockResponseDto
+                {
+                    Success = false,
+                    Message = "you are not administrator",
+                    Data = null,
+                });
             var response = await _repoService.DeleteRepo(id, expiry);
             if (response.success == false)
             {
@@ -93,6 +118,21 @@ namespace Elderly_Canteen.Controllers
         public async Task<ActionResult> RestockStuff(RestockRequestDto dto)
         {
             var adminId = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+            var identity = User.FindFirst(ClaimTypes.Role)?.Value;
+            if (adminId == null)
+                return BadRequest(new RestockResponseDto
+                {
+                    Success = false,
+                    Message = "not authoriented",
+                    Data = null,
+                });
+            if(identity != "admin")
+                return BadRequest(new RestockResponseDto
+                {
+                    Success = false,
+                    Message = "you are not administrator",
+                    Data = null,
+                });
             var response = await _repoService.Restock(dto, adminId);
             if(response.Success == true)
             {
@@ -107,7 +147,18 @@ namespace Elderly_Canteen.Controllers
         [HttpGet("search/restock")]
         public async Task<ActionResult> GetRestockHistory()
         {
-            return Ok();
+            var identity = User.FindFirst(ClaimTypes.Role)?.Value;
+            if (identity != "admin")
+                return BadRequest(new RestockResponseDto
+                {
+                    Success = false,
+                    Message = "you are not administrator",
+                    Data = null,
+                });
+            var response = await _repoService.GetRestockHistory();
+            if(response.Success != false)
+                return Ok(response);
+            return BadRequest(response);
         }
     }
 }
